Ensure map of RCTInspectorPackagerConnectionProtocol is cleared when React is destroyed
Summary:
There's a potential race between the static IInspector C++ instance being deallocated and the web socket connections in the RCTInspectorPackagerConnectionProtocol map receiving a message and attempting to call the static IInspector, leading to a crash in debug builds.

This ensures that the connections map is cleaned up with React, before the data segment cleanup occurs and destroys the static IInspector instance.

## Changelog

[IOS][Fixed] Fixes race condition causing crash in dev builds related to inspector protocol

Differential Revision: D63546862